Add marketing homepage and move chat
